Refactoring kernel to make space for a secondary public-kernel
